Fixing the bug with Pagination.
Basically what was happening was that when the user switched tabs from one that had more pages to one that had less, I needed to reset the page to the first one. Probably not the most complete solution, but the one I went with for this since it's just a simple app.